Message indexing
Sometimes messages are waiting to be sent for network interruption and the current logic tries to send them twice. Sometimes this leads to messages delivered multiple times - therefore I am now checking via an index if the message was previously delivered. This led also to change `useUpdateMessage` which was looping through all the messages: now we have an exact reference to the index of the message we need to update.